pyproject.toml
[project]
name = "llm-based-cluster-validation"
version = "0.1.0"
description = "DSPy-based cluster validation via intruder detection — internal demo and onboarding resource"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "bertopic>=0.17.4",
    "datasets>=4.8.4",
    "dspy-code[anthropic]>=0.1.5",
    "hdbscan>=0.8.42",
    "mlflow>=2.18.0",
    "pandas>=3.0.2",
    "peft>=0.18.1",
    "sentence-transformers>=5.3.0",
    "sglang[all]>=0.4.4.post3",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "trl>=1.0.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cluster_validator"]
